requirements.txt
# Core RAG dependencies
faiss-cpu==1.8.0
langchain-text-splitters==0.2.0
langchain-community==0.2.0

# CLI and utilities
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0

# Qdrant vector database client
qdrant-client>=1.7.0

# OpenAI API for embeddings
openai>=1.0.0